review: rename utils file from research-output to output
