+++ packages/gp2-components/src/utils/output-form.ts
